requirements.txt
Imath==0.0.2
numpy==1.24.4
OpenEXR==1.3.9
OpenEXR==1.3.9
Pillow==10.0.1
Pillow==10.1.0
torch==1.12.0+cu113
torchvision==0.13.0+cu113
torchaudio==0.12.0+cu113
scipy==1.10.1
tqdm==4.66.1